move modules and functions to diff files
