+++ launchcontainers/launchcontainers.py
@@ -1,4 +1,3 @@
-import nibabel as nib
 import os
 import subprocess as sp
 import numpy as np
