--- launchcontainers/packages_list.txt
@@ -1,8 +1,0 @@
-conda install dask
-conda install dask-jobqueue -c conda-forge
-conda install dask.distributed -c conda-forge 
-conda install shutil
-conda install pandas
-conda install subprocess
-conda install pyyaml
-conda install -c mrtrix3 mrtrix3
