+++ launchcontainers/prepare_input.py
@@ -63,4 +63,4 @@ def prepare_input_files(lc_config, lc_config_path, df_subSes, sub_ses_list_path,
                 "#####################################################\n")
     return new_lc_config_path, new_sub_ses_list_path,new_container_specific_config_path
 
-# %%
+
